@@ -30,13 +30,16 @@ class Banner extends Model
     protected $hidden = [];
 
     /**
-     * The attributes that should be cast.
+     * Get the attributes that should be cast.
      *
-     * @var array<string, string>
+     * @return array<string, string>
      */
-    protected $casts = [
-        'random_rotation' => 'boolean',
-    ];
+    protected function casts(): array
+    {
+        return [
+            'random_rotation' => 'boolean',
+        ];
+    }
 
     /**
      * Get the instance associated with the banner.
