@@ -34,11 +34,14 @@ class BannerConfiguration extends Model
     protected $hidden = [];
 
     /**
-     * The attributes that should be cast.
+     * Get the attributes that should be cast.
      *
-     * @var array<string, string>
+     * @return array<string, string>
      */
-    protected $casts = [];
+    protected function casts(): array
+    {
+        return [];
+    }
 
     /**
      * Get the banner template associated with the banner configuration.
