@@ -39,15 +39,18 @@ class BannerTemplate extends Model
     protected $hidden = [];
 
     /**
-     * The attributes that should be cast.
+     * Get the attributes that should be cast.
      *
-     * @var array<string, string>
+     * @return array<string, string>
      */
-    protected $casts = [
-        'enable_at' => 'datetime',
-        'disable_at' => 'datetime',
-        'enabled' => 'boolean',
-    ];
+    protected function casts(): array
+    {
+        return [
+            'enable_at' => 'datetime',
+            'disable_at' => 'datetime',
+            'enabled' => 'boolean',
+        ];
+    }
 
     /**
      * Get the banner associated with the model.
